feat: move SearchBox to more general component AppSearch
