+++ locales/de@informal.json
@@ -102,6 +102,7 @@
     "title": "Einstellungen"
   },
   "advancedSearch": {
+    "button": "Erweiterte Suche",
     "filters": {
       "email": "E-Mail",
       "firstname": "Vorname",
