+++ src/components/pages/admin/contacts/contacts.interface.ts
@@ -1,65 +1,6 @@
-import {
-  GetPaginatedQueryRuleOperator,
-  GetPaginatedQueryRuleValue,
-} from '../../../../utils/api/api.interface';
+import { Filters } from '../../../search/search.interface';
 
-export interface FilterArgs {
-  type: FilterType;
-  options?: string[];
-}
-
-export type FilterType =
-  | 'text'
-  | 'date'
-  | 'number'
-  | 'boolean'
-  | 'array'
-  | 'enum';
-
-export type FilterValue = GetPaginatedQueryRuleValue;
-
-export type FilterOperatorId = Exclude<
-  GetPaginatedQueryRuleOperator,
-  `${'is_not' | 'not'}_${string}`
->;
-
-export interface FilterOperatorParams {
-  args: number;
-}
-
-export const operators: Record<
-  FilterType,
-  Partial<Record<FilterOperatorId, FilterOperatorParams>>
-> = {
-  text: {
-    equal: { args: 1 },
-    contains: { args: 1 },
-    begins_with: { args: 1 },
-    ends_with: { args: 1 },
-  },
-  date: {
-    equal: { args: 1 },
-    between: { args: 2 },
-    less: { args: 1 },
-    greater: { args: 1 },
-  },
-  number: {
-    equal: { args: 1 },
-    less: { args: 1 },
-    greater: { args: 1 },
-  },
-  boolean: {
-    equal: { args: 1 },
-  },
-  array: {
-    contains: { args: 1 },
-  },
-  enum: {
-    equal: { args: 1 },
-  },
-};
-
-export const filters = {
+export const filters: Filters = {
   firstname: {
     type: 'text',
   },
@@ -111,29 +52,4 @@ export const filters = {
   tags: {
     type: 'array',
   },
-} as const;
-
-export type FilterId = keyof typeof filters;
-
-export interface EmptyFilter {
-  id: '';
-  operator: '';
-  inclusive: true;
-  values: [];
-}
-
-export interface Filter {
-  id: FilterId;
-  operator: FilterOperatorId;
-  inclusive: boolean;
-  values: FilterValue[];
-}
-export function emptyFilter(): EmptyFilter {
-  return { id: '', operator: '', inclusive: true, values: [] };
-}
-
-export function getOperators(
-  filter: Filter
-): Partial<Record<FilterOperatorId, FilterOperatorParams>> {
-  return operators[filters[filter.id].type];
-}
+};
