+++ src/components/search/operators/SingleArg.vue
@@ -3,7 +3,7 @@
   <ArgInput v-model="values[0]" :args="args" :readonly="readonly" />
 </template>
 <script lang="ts" setup>
-import { FilterArgs, FilterValue } from '../contacts.interface';
+import { FilterArgs, FilterValue } from '../search.interface';
 import ArgInput from './ArgInput.vue';
 defineProps<{ values: FilterValue[]; args: FilterArgs; readonly: boolean }>();
 </script>
